Fix HTML injection in column toggle.
The column toggle element gets the label as text an writes the label as HTML.

When using <script>alert("Hacked");</script> as column label (in HTML escaped form),
the code gets executed by the column toggle component.

As the column label might be user generated content, this could lead to XSS attacks.